+++ udata_recommendations/views.py
@@ -8,15 +8,9 @@ from udata.frontend import template_hook
 blueprint = Blueprint('recommendations', __name__, template_folder='templates')
 
 
-NB_RECOMMENDATIONS = 2
-
-
 def has_recommendations(ctx):
     dataset = ctx['dataset']
-    if not dataset:
-        return False
-    recommendations = dataset.extras.get('recommendations', [])
-    return recommendations and len(recommendations) > 0
+    return dataset and dataset.extras.get('recommendations', [])
 
 
 @template_hook('dataset.display.after-description', when=has_recommendations)
@@ -26,13 +20,8 @@ def dataset_recommendations(ctx):
 
     # Get at most n unique recommendations
     # Recommendations are already sorted by score in desc order
-    while len(reco_ids) < current_app.config.get('RECOMMENDATIONS_NB_RECOMMENDATIONS'):
-        try:
-            recommendation = recommendations.pop(0)
-            if recommendation['id'] not in reco_ids:
-                reco_ids.append(recommendation['id'])
-        except IndexError:
-            break
+    limit = current_app.config.get('RECOMMENDATIONS_NB_RECOMMENDATIONS')
+    reco_ids = list({r['id']: 0 for r in recommendations})[:limit]
 
     return theme.render(
         'dataset-recommendations.html',
